+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Please prefix each entry with one of:
 
 
 # Next Release
+- [fix] Assembly patterns could not match on full assembly name when using `@AssemblyDir` as PDB directory
 
 # v24.5.1
 - [fix] Upload Daemon uploaded to Artifactory in the wrong format